+++ app/Main.hs
@@ -9,13 +9,9 @@ main :: IO ()
 main = do
   resolution <- getScreenSize
 
-  arrow     <- loadSprite "./images/"  "right_arrow.png"
-  backgound <- loadSprite "./images/" "background"
-  cactus    <- loadSprite "./images/" "cactus"
-  clouds    <- loadSprite "./images/" "clouds"
-  finish    <- loadSprite "./images/" "finish"
+  assets <- sequence (loadSprite imageDir <$> sprites)
 
-  veh_suv   <- loadAnimation "./images/suv/" "veh_suv" 32
+  veh_suv <- loadAnimation "./images/suv/" "veh_suv" 32
 
   fontWaffle <- loadFont
     "./images/fontWaffle/"
@@ -24,6 +20,9 @@ main = do
 
   outrunPlay
     resolution
-    [backgound, clouds, veh_suv]
+    (veh_suv : assets)
     (proccessFontColors [afr32_olive, afr32_cyan, afr32_turmeric] fontWaffle)
-    (desertTrack [arrow, cactus, finish])
+    (desertTrack assets)
+  where
+    imageDir = "./images/"
+    sprites  = ["background", "clouds", "cactus", "finish", "right_arrow"]