+++ src/Outrun/Data/Defaults.hs
@@ -28,4 +28,4 @@ defaultCamera =
   Camera
     (0, 350, 0)
     defaultWindowResolution
-    0.325 100 0.8 0
+    0.45 200 0.85 0